+++ builder_derive/Cargo.toml
@@ -2,7 +2,7 @@
 name = "builder_derive"
 version = "0.1.1"
 authors = ["Andrew Weiss <wvvwwvw@gmail.com>"]
-edition = "2018"
+edition = "2021"
 
 [dependencies]
 syn = { version = "0.15.14", features = ["full", "extra-traits"]}
