+++ jvm_bare_agent/Cargo.toml
@@ -2,7 +2,7 @@
 name = "rollbar-jvm-agent"
 version = "0.1.1"
 authors = ["Andrew Weiss <wvvwwvw@gmail.com>"]
-edition = "2018"
+edition = "2021"
 
 [lib]
 crate_type = [ "cdylib" ]
