+++ jvm_core/Cargo.toml
@@ -3,7 +3,7 @@ name = "rollbar-jvm"
 version = "0.1.1"
 authors = ["Andrew Weiss <wvvwwvw@gmail.com>"]
 build = "build.rs"
-edition = "2018"
+edition = "2021"
 
 [dependencies]
 log = "0.4"
